pbs-264--feat: Header 리팩토링 완료
