+++ frontend/src/assets/css/Header.css
@@ -1,111 +1,168 @@
-.Container {
+/******************************************************************/
+/* 헤더 ***********************************************************/
+/******************************************************************/
+.Header_Container > div {
     width: 100%;
-    height: 84px;
-    display: grid;
+    height: 80px;
+    margin: 0 auto 0 auto;
     box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    padding: 0 0 8px 0;
+    font-family: 'Pretendard-Regular';
+    font-weight: 600;
 }
 
-.Container > div {
-    display: grid;
-    /*grid-template-columns: 45px 340px 170px 144px 150px 150px 44px 166px 55px;*/
-    /*grid-template-columns: 0.5fr 1.6fr 1.5fr 1.3fr 1.6fr 1.4fr 0.4fr 1.5fr 0.5fr; 알림 컴포넌트가 가려서 결국 px로.. */
-    justify-content: center;
-    align-items: center;
-    gap: 10px;
-}
-
-/* POSCO 로고 */
-.Container > div > img {
-    width: 120px;
-}
-
-/* 메뉴 링크 */
-.Container > div > div:nth-child(3),
-.Container > div > div:nth-child(5),
-.Container > div > div:nth-child(4) {
-    font-size: 1.5em;
-    font-weight: bold;
-    color: #03507d;
-}
-
-/* 로그인 아이콘 + 사용자 이름 */
-.Container > div > div:nth-child(6) {
+.Header_Container > div > div {
     display: flex;
-    width: fit-content;
-    font-size: 1.2em;
-    font-weight: bold;
-    color: #03507d;
-    align-items: center;
 }
 
-/* 로그인 아이콘 */
-.Container > div > div:nth-child(6) img {
-    width: 36px;
-    margin: 0 0.5vw 0 0;
+/* 상단 섹션 */
+.Header_Container > div > div:first-child {
+    width: 1320px;
+    align-content: center;
+    margin: 8px auto 0 auto;
+    font-size: 14px;
 }
 
-/* 알림 버튼 */
-.Container > div > div:nth-child(7) Button:hover {
-    background-color: #03507d !important;
-    color: #ffffff !important;
-    border: none !important;
-}
-
-/* 로그아웃 버튼 */
-.Container > div > div:nth-child(8) Button:hover {
-    background-color: pink !important;
-    background-color: #03507d !important;
-    color: #ffffff !important;
-    border: none !important;
-}
-
-.Container_Nav {
-    display: grid;
+.Header_Container > div hr {
     width: 100%;
-    height: max-content;
+    border: 1px solid #d5dbe2;
+}
+
+/* 하단 섹션 */
+.Header_Container > div > div:nth-child(3) {
+    width: 1320px;
+    margin: 0 auto 0 auto;
+}
+
+/* pobluesky 로고 + 상세 메뉴 + 프로필 사진 + 회원명 */
+.Header_Container > div > div:nth-child(3) > div:nth-child(1),
+.Header_Container > div > div:nth-child(3) > div:nth-child(2) {
+    display: flex;
+    align-items: center;
+    font-size: 18px;
+}
+
+/* pobluesky 로고 */
+.Header_Container
+    > div
+    > div:nth-child(3)
+    > div:nth-child(1)
+    > div:nth-child(1)
+    img {
+    width: 96px;
+    cursor: pointer;
+    margin: 0 36px 0 0;
+}
+
+/* 프로필 사진 + 회원명 */
+.Header_Container > div > div:nth-child(3) > div:nth-child(2) {
+    margin: 0 0 0 auto;
+}
+
+/* 프로필 사진 */
+.Header_Container
+    > div
+    > div:nth-child(3)
+    > div:nth-child(2)
+    > div:nth-child(1)
+    img {
+    height: 36px;
+    margin: 4px 0 0 0;
+}
+
+/* 프로필 사진, 회원명 간 간격 */
+.Header_Container > div > div:nth-child(3) > div:nth-child(2) > :nth-child(n) {
+    margin: 0 0 0 24px;
+}
+
+.Header_Container button {
+    background-color: transparent;
+    border: none;
+    cursor: pointer;
+}
+
+/******************************************************************/
+/* 사용자 정보 모달 ************************************************/
+/******************************************************************/
+.User_Modal_Container {
     position: absolute;
+    width: 1320px;
+    top: 5px;
+    left: 50%;
+    transform: translate(-50%);
+    z-index: 1;
+}
+
+.User_Modal_Container > div {
+    position: relative;
+    width: 196px;
+    padding: 20px;
+    height: fit-content;
+    border: 1px solid #d5dbe2;
+    border-radius: 8px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    background-color: #ffffff;
+    float: right;
+}
+
+.User_Modal_Container > div Button:hover {
+    border: 1px solid #25262b !important;
+    color: #25262b !important;
+    font-weight: 600;
+}
+
+/******************************************************************/
+/* 알림 모달 *******************************************************/
+/******************************************************************/
+.Notify_Modal_Container {
+    position: absolute;
+    width: 1320px;
+    top: 5px;
+    left: 50%;
+    transform: translate(-50%);
+    z-index: 1;
+}
+
+.Notify_Modal_Container > div {
+    width: 300px;
+    float: right;
+    padding: 20px;
+    border: 1px solid #d5dbe2;
+    border-radius: 10px;
+    background: white;
     box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
 }
 
-.Container_Nav > div {
-    display: grid;
-    grid-template-columns: 45px 340px 170px 144px 150px 150px 44px 166px 55px;
-    background-color: #ffffff;
-    position: relative;
-    justify-content: center;
-    gap: 10px;
+/* 새 알림, 읽은 알림 */
+.Notify_Modal_Container > div > div:first-child {
+    display: flex;
+    margin: 0 0 20px 0;
 }
 
-.Container_Nav > div > div:nth-child(2) {
-    width: 120px;
+/* 알림 목록 그룹 */
+.Notify_Modal_Container > div > div:nth-child(2) > div:first-child {
+    height: 100%;
+    overflow-y: auto;
 }
 
-/* Inquiry 메뉴 */
-.Container_Nav > div > div:nth-child(3) {
-    margin: 2vh 0 2vh 0;
+.Notify_Modal_Container
+    > div
+    > div:nth-child(2)
+    > div:first-child
+    > div
+    > div:nth-child(2) {
+    font-size: 15px;
+    line-height: 20px;
+    font-weight: 700;
+    text-align: left;
 }
 
-/* Inquiry 메뉴 리스트별 */
-.Container_Nav > div > div:nth-child(3) > div {
-    margin: 2vh 0 2vh 0;
-}
-
-/* Inquiry 메뉴 리스트별 */
-.Container_Nav > div > div:nth-child(3) > div:hover {
-    font-weight: 600;
-}
-
-/* VoC 메뉴 */
-.Container_Nav > div > div:nth-child(4) {
-    margin: 2vh 0 2vh 0;
-}
-
-/* VoC 메뉴 리스트별 */
-.Container_Nav > div > div:nth-child(4) > div {
-    margin: 2vh 0 2vh 0;
-}
-
-/* VoC 메뉴 리스트별 */
-.Container_Nav > div > div:nth-child(4) > div:hover {
-    font-weight: 600;
+.Notify_Modal_Container
+    > div
+    > div:nth-child(2)
+    > div:first-child
+    > div
+    > div:nth-child(2)
+    span {
+    font-weight: 400;
 }
