+++ frontend/src/components/molecules/UserInfoModal.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import Button from '../atoms/Button';
 import { useAuth } from '../../hooks/useAuth';
-import { Modal_Container } from '../../assets/css/MyHeader.css';
+import { User_Modal_Container } from '../../assets/css/Header.css';
 
 // 로그아웃 버튼, 회원 정보 탐색 기능
 function UserInfoModal() {
@@ -15,36 +15,33 @@ function UserInfoModal() {
     };
 
     return (
-        <div className={Modal_Container}>
+        <div className={User_Modal_Container}>
             <div>
-                <div>
-                    <Button
-                        onClick={() => {
-                            alert('내 정보 페이지 구현 중');
-                        }}
-                        btnName={'내 정보'}
-                        width={'100%'}
-                        height={'40px'}
-                        backgroundColor={'#ffffff'}
-                        textColor={'#64636A'}
-                        border={'solid #c1c1c1 1px'}
-                        borderRadius={'12px'}
-                        fontSize={'16px'}
-                    />
-                </div>
-                <div>
-                    <Button
-                        onClick={() => handleLogout()}
-                        btnName={'로그아웃'}
-                        width={'100%'}
-                        height={'40px'}
-                            backgroundColor={'#ffffff'}
-                        textColor={'#64636A'}
-                        border={'solid #c1c1c1 1px'}
-                        borderRadius={'12px'}
-                        fontSize={'16px'}
-                    />
-                </div>
+                <Button
+                    onClick={() => {
+                        alert('내 정보 페이지 구현 중');
+                    }}
+                    btnName={'내 정보 수정'}
+                    width={'fit-content'}
+                    height={'40px'}
+                    backgroundColor={'#ffffff'}
+                    textColor={'#64636a'}
+                    border={'solid #64636a 1px'}
+                    borderRadius={'12px'}
+                    fontSize={'16px'}
+                />
+                <Button
+                    onClick={() => handleLogout()}
+                    btnName={'로그아웃'}
+                    width={'fit-content'}
+                    height={'40px'}
+                    backgroundColor={'#ffffff'}
+                    textColor={'#d5dbe2'}
+                    border={'solid #d5dbe2 1px'}
+                    borderRadius={'12px'}
+                    fontSize={'16px'}
+                    float={'right'}
+                />
             </div>
         </div>
     );
