+++ frontend/src/index.css
@@ -27,6 +27,10 @@ input:focus {
     outline: none;
 }
 
+button {
+    outline: none;
+}
+
 @font-face {
     font-family: 'Pretendard-Regular';
     src: url('https://fastly.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Regular.woff')
